# access log: speed up the access log formatter