*Description*:
This change reduces the CPU usage of the access log formatter ~10% by using `fmt::FormatInt` in place of the much slower `std::to_string`.

Benchmark results: 
```
-------------------------------------------------------------
Benchmark                      Time           CPU Iterations
-------------------------------------------------------------
Before:
BM_AccessLogFormatter       5401 ns       5401 ns     129597

After:
BM_AccessLogFormatter       4788 ns       4788 ns     146003
```

*Risk Level*: Medium
*Testing*: unit tests
*Docs Changes*: N/A
*Release Notes*: N/A
*Issues*: #2879

Signed-off-by: Brian Pane <bpane@pinterest.com>